move abstractions to standalone module
